third_party/blink/web_tests/external/wpt/custom-elements/registries/ShadowRoot-init-customElementRegistry-expected.txt

This is a testharness.js-based test.
[FAIL] A newly attached disconnected ShadowRoot should use the global registry by default
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] A newly attached connected ShadowRoot should use the global registry by default
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
[FAIL] A newly attached disconnected ShadowRoot should use the scoped registry if explicitly specified in attachShadow
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] A newly attached connected ShadowRoot should use the scoped registry if explicitly specified in attachShadow
  Failed to construct 'CustomElementRegistry': Illegal constructor
[FAIL] attachShadow() should use the global registry when customElementRegistry is null
  assert_equals: expected (object) object "[object CustomElementRegistry]" but got (undefined) undefined
Harness: the test ran to completion.
